% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{IB.areachange.pol}
\alias{IB.areachange.pol}
\title{IB.areachange.pol creates polygon shapefiles per archipelago at each sea level stand from a sea level curve and a bathymetry DEM. It attaches a name to each polygon based on the names provided by the user in a separate KML file.}
\usage{
IB.areachange.pol(IB.Arcname, IB.projname, IB.proj, IB.inputfolder,
  IB.intermedfolder, IB.outputfolder, time.ka, time.sl)
}
\arguments{
\item{IB.Arcname, }{the name of the archipelago}

\item{IB.projname, }{name of the archipelago bathymetry DEM which was produced in IB.WMcrop.proj.r (without name of archipelago, and without file extention).}

\item{IB.proj, }{coordinate reference system (crs) to which the shapefiles should be projected}

\item{IB.inputfolder, }{file path to input folder}

\item{IB.intermedfolder, }{file path to intermediate folder}

\item{IB.outputfolder, }{file path to output folder}

\item{time.ka, }{vector with time periods corresponding to time.sl}

\item{time.sl, }{vector with sea level stands corresponding to time.ka}
}
\description{
IB.areachange.pol creates polygon shapefiles per archipelago at each sea level stand from a sea level curve and a bathymetry DEM. It attaches a name to each polygon based on the names provided by the user in a separate KML file.
}
\author{
Sietze Norder
}
