% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{IB.areamerge}
\alias{IB.areamerge}
\title{IB.areamerge combines the csv files of separate archipelagos into a single file.}
\usage{
IB.areamerge(IB.Arclist, IB.outputfolder)
}
\arguments{
\item{IB.Arclist, }{list with names of archipelagos}

\item{IB.outputfolder, }{file path to output folder}
}
\description{
IB.areamerge combines the csv files of separate archipelagos into a single file.
}
\author{
Sietze Norder
}
