% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{IB.datastructure}
\alias{IB.datastructure}
\title{IB.datastructure selects island names from KML files per archipelago and stores island names of all archipelagos in a csv file. This is just to provide an overview of which islands are contained in the dataset.}
\usage{
IB.datastructure(IB.inputfolder, IB.Arclist)
}
\arguments{
\item{IB.inputfolder, }{file path to input folder}

\item{IB.Arclist, }{list with names of archipelagos}
}
\description{
IB.datastructure selects island names from KML files per archipelago and stores island names of all archipelagos in a csv file. This is just to provide an overview of which islands are contained in the dataset.
}
\author{
Sietze Norder
}
