% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{IB.projpol}
\alias{IB.projpol}
\title{IB.projpol projects shapefiles of t=0 to a user-defined projection and saves it as new shapefile:}
\usage{
IB.projpol(IB.inputfolder, IB.outputfolder, IB.Arclist, IB.projcrs)
}
\arguments{
\item{IB.Arclist, }{list with names of archipelagos}

\item{IB.projcrs, }{crs to which the shapefile should be projected.}

\item{IB.intermedfolder, }{file path to intermediate folder}
}
\description{
IB.projpol projects shapefiles of t=0 to a user-defined projection and saves it as new shapefile:
}
\author{
Sietze Norder
}
