% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{IB.WMcrop.proj.r}
\alias{IB.WMcrop.proj.r}
\title{IB.WMcrop.proj.r crops the worldmap to an extent which is slightly larger than the bounding box of an archipelago. It also reprojects the resulting raster to a user-defined crs.}
\usage{
IB.WMcrop.proj.r(IB.Arcname, IB.projname, IB.proj, IB.WM, resfact,
  IB.inputfolder, IB.outputfolder, IB.plot = TRUE, IB.mask = FALSE)
}
\arguments{
\item{IB.Arcname, }{the name of the archipelago}

\item{IB.projname, }{name of the archipelago bathymetry DEM to be stored (without name of archipelago, and without file extention). The name provided here should correspond to the name provided in the function IB.areachange.pol}

\item{IB.proj, }{coordinate reference system (crs) to which the raster should be projected}

\item{IB.WM, }{file path where global bathymetry DEM is stored locally.}

\item{resfact, }{amount of disaggregation (expressed as number of cells) to create a raster in higher resolution}

\item{IB.inputfolder, }{file path to input folder}

\item{IB.outputfolder, }{file path to output folder}

\item{IB.plot, }{should output be plotted on screen?}

\item{IB.mask, }{if TRUE: assign NA to areas below sea-level of -150m MSL (this reduces computation time)}
}
\description{
IB.WMcrop.proj.r crops the worldmap to an extent which is slightly larger than the bounding box of an archipelago. It also reprojects the resulting raster to a user-defined crs.
}
\author{
Sietze Norder
}
