% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAC_functions.R
\name{polygonizer}
\alias{polygonizer}
\title{Polygonizer creates a polygon shapefile from a raster layer}
\usage{
polygonizer(x, outshape = NULL, pypath = NULL, readpoly = TRUE,
  fillholes = FALSE, aggregate = FALSE, quietish = TRUE)
}
\arguments{
\item{x:}{an R Raster layer, or the file path to a raster file recognised by GDAL}

\item{outshape:}{the path to the output shapefile (if NULL, a temporary file will
be created)}

\item{pypath:}{the path to gdal_polygonize.py or OSGeo4W.bat (if NULL, the function
will attempt to determine the location). OSGeo4W can be downloaded from
https://trac.osgeo.org/osgeo4w/}

\item{readpoly:}{should the polygon shapefile be read back into R, and returned by
this function? (logical)}

\item{fillholes:}{should holes be deleted (i.e., their area added to the containing
polygon)}

\item{aggregate:}{should polygons be aggregated by their associated raster value?}

\item{quietish:}{should (some) messages be suppressed? (logical)}
}
\description{
Polygonizer creates a polygon shapefile from a raster layer
}
\author{
Sietze Norder
}
